+++ src/main/java/org/cana/services/gameplayer/repository/GamePlayerRepository.java
@@ -4,14 +4,12 @@ import io.quarkus.hibernate.orm.panache.PanacheRepository;
 import org.cana.services.gameplayer.repository.daos.GamePlayer;
 
 import javax.enterprise.context.ApplicationScoped;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 @ApplicationScoped
 public class GamePlayerRepository implements PanacheRepository<GamePlayer>
 {
-    public List<GamePlayer> findByIdAndIsActive(int gameId)
+    public List<GamePlayer> findByIdAndIsActive(Long gameId)
     {
         return  list("gameid= ?1 and isactive=true",gameId);
     }
